Remove unbounded check for "install.php" in `wp_not_installed()`.
The check is true for `wp-admin/plugin-install.php` and `wp-admin/theme-install.php` too. This will end in a redirect to `/wp-admin/upgrade.php` or a screen with a bunch of errors.
The `WP_INSTALLING` constant was added in WordPress 1.5.1 to `wp-admin/install.php` which makes the check for "install.php" redundant.

props joemcgill.
fixes #29694.
Built from https://develop.svn.wordpress.org/trunk@30581