List tables: Ensure special CSS for the title column gets applied.
This could have some side effects if a custom list table has a title column with a strong element inside that is not the post title, but that is fairly edge and we can address that if it comes up. Also moves the rules into `list-tables.css`.

see #25408.

Built from https://develop.svn.wordpress.org/trunk@33105