Accessibility: Remove the title attribute from the Featured Image postbox thumbnail.
The title attribute is now replaced with plain text in a `howto` paragraph, as done for example in the Tags postbox.
Adds an `aria-describedby` attribute on the thumbnail link to target the description.
Also, fixes the focus style on the featured image thumbnail.

Fixes #35076.
Built from https://develop.svn.wordpress.org/trunk@36000